Make user manager optional Split ChatFirestore into several files Update typing completion block to return only user ids
